Made SentenceTokenizer into a separate class.
